Use conditional compilation for url capturing per platform
